Create NAV2-1.1.1a
Using ARIA features helps to define different sections of a page, to indicate the currently active page within a set of pagination links, to define the total number of items in a set, and to define the position of an item within a set. This enables the screen reader to navigate through paginated contents more accurately.